New classes: expressionfolder.py globs a folder and uses a regular expression to choose wanted files ie. images. Replaces imagefolder.py
imageobject.py is a wrapper for image filename data. Can use a regular expression in order to get
information from filename.

mock_stdic.py is a mock version of the whole program, written to illustrate the course program will take.
Will be updated constantly as different classes are named and added. Once completed, can be used as a basis for
working release.